Adds port connection and implicite library usage
Single ports of an instance can now be connected to the entity and to a
signal of the architecture.  In case the signal is not part of the
architecture it is created.  If an entity or architecture uses a
generic, port or signal with a known data type the library is added to
the entity.